backend/requirements.txt
flask_cors
flask
flask-sqlalchemy
google.generativeai
firebase
google_auth_oauthlib
pymysql
filetype
gunicorn